Move clients to rtk folder
