+++ packages/grafana-api-clients/src/clients/rtkq/advisor/v0alpha1/baseAPI.ts
@@ -1,7 +1,7 @@
 import { createApi } from '@reduxjs/toolkit/query/react';
 
-import { createBaseQuery } from '../../../utils/createBaseQuery';
-import { getAPIBaseURL } from '../../../utils/utils';
+import { createBaseQuery } from '../../../../utils/createBaseQuery';
+import { getAPIBaseURL } from '../../../../utils/utils';
 
 export const API_GROUP = 'advisor.grafana.app' as const;
 export const API_VERSION = 'v0alpha1' as const;
