+++ packages/grafana-api-clients/src/clients/rtkq/dashboard/v0alpha1/index.ts
