+++ packages/grafana-api-clients/src/clients/rtkq/preferences/user/baseAPI.ts
@@ -1,9 +1,9 @@
 import { createApi } from '@reduxjs/toolkit/query/react';
 
-import { createBaseQuery } from '../../utils/createBaseQuery';
+import { createBaseQuery } from '../../../../utils/createBaseQuery';
 
 export const api = createApi({
-  reducerPath: 'migrateToCloudGeneratedAPI',
+  reducerPath: 'userPreferencesAPI',
   baseQuery: createBaseQuery({ baseURL: '/api' }),
   endpoints: () => ({}),
 });
