+++ packages/grafana-api-clients/src/clients/rtkq/preferences/user/index.ts
