+++ packages/grafana-api-clients/src/clients/rtkq/preferences/v1alpha1/endpoints.gen.ts
