+++ packages/grafana-api-clients/src/clients/rtkq/shorturl/v1alpha1/endpoints.gen.ts
